Refacor OTXBenchmark
